Add code for hardware test (not called from main yet)
