+++ Software/Testing/micropython/software/wav_utils.py
@@ -1,0 +1,18 @@
+def create_wav_header(sampleRate, bitsPerSample, num_channels, num_samples):
+    datasize = num_samples * num_channels * bitsPerSample // 8
+    o = bytes("RIFF", "ascii")  # (4byte) Marks file as RIFF
+    o += (datasize + 36).to_bytes(
+        4, "little"
+    )  # (4byte) File size in bytes excluding this and RIFF marker
+    o += bytes("WAVE", "ascii")  # (4byte) File type
+    o += bytes("fmt ", "ascii")  # (4byte) Format Chunk Marker
+    o += (16).to_bytes(4, "little")  # (4byte) Length of above format data
+    o += (1).to_bytes(2, "little")  # (2byte) Format type (1 - PCM)
+    o += (num_channels).to_bytes(2, "little")  # (2byte)
+    o += (sampleRate).to_bytes(4, "little")  # (4byte)
+    o += (sampleRate * num_channels * bitsPerSample // 8).to_bytes(4, "little")  # (4byte)
+    o += (num_channels * bitsPerSample // 8).to_bytes(2, "little")  # (2byte)
+    o += (bitsPerSample).to_bytes(2, "little")  # (2byte)
+    o += bytes("data", "ascii")  # (4byte) Data Chunk Marker
+    o += (datasize).to_bytes(4, "little")  # (4byte) Data size in bytes
+    return o
